requirements.txt
python-telegram-bot==20.3
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
